Extracted BufferCacheKey
